+++ Modules/Ruleset.ProgramInfo/Public/Search-Installation.ps1
@@ -554,6 +554,10 @@ function Search-Installation
 		{
 			Update-Table -Search "Brave"
 		}
+		"nmap"
+		{
+			Update-Table -Search "Nmap"
+		}
 		"Chrome"
 		{
 			Update-Table -Search "Google Chrome" -UserProfile